requirements.txt
# SPDX-FileCopyrightText: 2022 Alec Delaney, written for Adafruit Industries
#
# SPDX-License-Identifier: MIT

parse
pygithub
GitPython
typer
typing-extensions
